pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "desert-patterns"
version = "0.1.0"
description = "Unsupervised discovery of patterned ground types in Atacama Desert imagery"
requires-python = ">=3.10"
dependencies = [
    "exifread>=3.0",
    "Pillow>=10.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "pandas>=2.0",
    "PyYAML>=6.0",
    "scikit-image>=0.21",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "h5py>=3.9",
    "umap-learn>=0.5",
    "hdbscan>=0.8.30",
    "joblib>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",
]

[project.optional-dependencies]
viz = [
    "gradio>=4.0",
]
ml = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.40",
    "segment-anything @ git+https://github.com/facebookresearch/segment-anything.git",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 99
target-version = "py310"
